+++ src/ArrayMicRefreshment.App/SettingsCopier.cs
@@ -21,6 +21,7 @@ internal static class SettingsCopier
         target.WakeCommandSilenceMs = source.WakeCommandSilenceMs;
         target.WakeUseVadEndDetection = source.WakeUseVadEndDetection;
         target.HudScreenCorner = source.HudScreenCorner;
+        target.UseWebStatusHud = source.UseWebStatusHud;
         target.SkillsDirectory = source.SkillsDirectory;
         target.ModelsDirectory = source.ModelsDirectory;
         target.SelectedAsrModelId = source.SelectedAsrModelId;